pyproject.toml
[project]
name = "newproject"
version = "0.1.0"
description = ""
authors = [
    {name = "Shkaf4i3",email = "fleximes@protonmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all] (>=0.120.1,<0.130.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]


[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
